backend/metricMapAPI/metrics/management/future_features/backup/dashboard/scenario_simulation.py
"""
This module is designed to simulate data under various hypothetical scenarios by adjusting parameters like growth rate,
noise level, and seasonality. It allows for flexible scenario analysis to understand potential impacts on data trends.
"""

import numpy as np
import matplotlib.pyplot as plt

def simulate_scenario(base_data, scenario_params):
    """
    Functions:
    simulate_scenario(base_data, scenario_params): Modifies a dataset according to specified scenario parameters
        such as growth rate, noise level, and seasonality to simulate various business or environmental conditions.
    """
    simulated_data = base_data.copy()
    
    if 'growth_rate' in scenario_params:
        growth_rate = scenario_params['growth_rate']
        simulated_data['value'] *= (1 + growth_rate)
    
    if 'noise_level' in scenario_params:
        noise_level = scenario_params['noise_level']
        noise = np.random.normal(0, noise_level, len(simulated_data))
        simulated_data['value'] += noise
    
    if 'seasonality' in scenario_params:
        seasonality = scenario_params['seasonality']
        simulated_data['value'] += seasonality * np.sin(np.arange(len(simulated_data)) * (2 * np.pi / 365))
    
    return simulated_data

"""
Example:
    >>> base_data = pd.DataFrame({
        'date': pd.date_range(start='2021-01-01', periods=365, freq='D'),
        'value': np.linspace(100, 200, 365)
    })
    >>> scenario_params = {'growth_rate': 0.05, 'noise_level': 1, 'seasonality': 20}
    >>> simulated_data = simulate_scenario(base_data, scenario_params)
    >>> plt.plot(simulated_data['date'], simulated_data['value'])
    >>> plt.show()

Notes:
    - The function returns a DataFrame containing the simulated data.
    - It uses numpy for numerical operations and optionally matplotlib for visualizing the results if desired.
    - Parameters for simulation include:
        'growth_rate': Percentage increase or decrease applied uniformly across the dataset.
        'noise_level': Standard deviation of normally distributed random noise added to the data.
        'seasonality': Amplitude of sinusoidal variation to model seasonal effects.
    - The scenario parameters are optional and can be combined as needed to reflect complex real-world conditions.
"""